+++ package.json
@@ -13,7 +13,8 @@
     "ejs": "^2.5.5",
     "express": "^4.14.0",
     "express-handlebars": "^3.0.0",
-    "jade": "^1.11.0"
+    "jade": "^1.11.0",
+    "mongodb": "^2.2.12"
   },
   "devDependencies": {
     "gulp": "^3.9.1",
