+++ src/views/bookListView.ejs
@@ -117,7 +117,7 @@
                                 <div class="col-xs-9">
                                     <h4><span class="label label-default"><%= books[i].author %></span></h4>
                                     <h4>
-                                        <a href="/books/<%= i %>" class="text-muted">Read More</a></small>
+                                        <a href="/books/<%= books[i]._id %>" class="text-muted">Read More</a></small>
                                     </h4>
                                 </div>
                                 <div class="col-xs-3"></div>
